chatbot_backend_FastAPI/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dataclasses-json==0.6.7
DateTime==5.5
fastapi==0.115.12
frozenlist==1.7.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
jwt==1.4.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-text-splitters==0.3.8
langsmith==0.4.4
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.1
orjson==3.10.18
packaging==24.2
passlib==1.7.4
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.41
SQLAlchemy-Utils==0.41.2
starlette==0.46.2
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.5.0
uvicorn==0.34.2
yarl==1.20.1
zope.interface==7.2
zstandard==0.23.0
